[softwaremanager] prevent crash and check if images is present

+++ lib/python/Plugins/SystemPlugins/SoftwareManager/plugin.py
@@ -1734,7 +1734,9 @@ class UpdatePlugin(Screen):
 				if not os.path.exists("/media/hdd"):
 					print "no hdd present to save xionrestore file"
 				else:
-					open('/media/hdd/images/xionrestore','w').close()
+					if not os.path.exists("/media/hdd/images"):
+						os.makedirs('/media/hdd/images')
+						open('/media/hdd/images/xionrestore','w').close()
 			else:
 				print "Xion skin update = No"
 				if os.path.exists('/media/hdd/images/hdfrestore'):
